refactor: rename trimmed results.py to report.py
