+++ pyproject.toml
@@ -99,7 +99,6 @@ ban-relative-imports = "all" # enforce absolute imports (TID252)
 # only restate it, so we don't require a docstring on every override.
 "src/delta_engine/domain/plan/actions.py" = ["D102"]
 "src/delta_engine/application/failures.py" = ["D102"]
-"src/delta_engine/application/results.py" = ["D102"]
 # Example notebooks in Databricks source format. Databricks injects `spark`,
 # `dbutils`, and `display` at runtime; the first line must be the notebook
 # header (so no module docstring); imports follow cell markers, not the file top.
