+++ tests/application/test_public_api.py
@@ -16,9 +16,9 @@ def test_public_api_exposes_the_intended_names():
     )
     from delta_engine.application.engine import Engine as EngineImpl
     from delta_engine.application.errors import SyncFailedError as SyncFailedErrorImpl
+    from delta_engine.application.failures import Failure as FailureImpl
     from delta_engine.application.registry import Registry as RegistryImpl
-    from delta_engine.application.results import (
-        Failure as FailureImpl,
+    from delta_engine.application.report import (
         SyncReport as SyncReportImpl,
         TableRunStatus as TableRunStatusImpl,
     )
